ft: cleanup API, distinguish outbound/inbound connections, redirect unauthorized settings to login page
