+++ api/app.py
@@ -14,8 +14,10 @@ from api import API_ROOT, basic_auth, bcrypt, db, login_manager
 from api.admin.actions import actions_bp
 from api.admin.APIAuth import APIAuth
 from api.admin.APIConfig import APIConfig
-from api.admin.model import AuthAdmin, Connection, ConnectionAdmin
+from api.admin.inbound_connection import InboundConnection, InboundConnectionView
 from api.admin.nodes import nodes_bp
+from api.admin.outbound_connection import OutboundConnection, OutboundConnectionView
+from api.admin.user_account_control import UserAccountControlView
 from api.admin.views import Logout, SettingsView
 from api.user.author.model import Author
 from api.user.comments.model import Comment
@@ -42,9 +44,11 @@ def create_app(testing_env=False):
         # otherwise, we should be serving a frontend resource here
         return send_from_directory(app.static_folder, path)
 
+    ADMIN_ENDPOINT = f"{API_ROOT}/admin/action"
+
     app.register_blueprint(user_bp, url_prefix=f"{API_ROOT}/authors")
     app.register_blueprint(nodes_bp, url_prefix=f"{API_ROOT}/nodes")
-    app.register_blueprint(actions_bp, url_prefix=f"{API_ROOT}/admin/action")
+    app.register_blueprint(actions_bp, url_prefix=ADMIN_ENDPOINT)
 
     app.config.from_object("api.config.Config")
 
@@ -58,11 +62,20 @@ def create_app(testing_env=False):
 
     # admin views
     admin = Admin(app, name="bigger-yoshi", template_mode="bootstrap3")
-    admin.add_view(AuthAdmin(Author, db.session))
-    admin.add_view(ConnectionAdmin(Connection, db.session))
     admin.add_view(SettingsView(name="Settings", endpoint="settings"))
+    admin.add_view(UserAccountControlView(Author, db.session, name="User Account Control", endpoint="accounts"))
+    admin.add_view(
+        OutboundConnectionView(
+            OutboundConnection, db.session, name="Outbound server connections", endpoint="connections"
+        )
+    )
+    admin.add_view(
+        InboundConnectionView(
+            InboundConnection, db.session, name="Inbound server connections", endpoint="connections_inbound"
+        )
+    )
     admin.add_view(Logout(name="logout", endpoint="Logout"))
-    app.jinja_env.globals.update(APIConfig=APIConfig)
+    app.jinja_env.globals.update(APIConfig=APIConfig, admin_endpoint=ADMIN_ENDPOINT)
 
     # docs
     p = Path(__file__).with_name("swagger.json")
