+++ api/templates/not_admin_page.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Admin page</title>
+</head>
+<body>
+<h1>{{ current_username }} does not have admin privileges.</h1>
+<h2>Please contact your domain administrator to arrange access.</h2>
+</body>
+</html>
